Speed up PostControllerTest
